separate aot-runtime
