+++ pom.xml
@@ -36,6 +36,7 @@
   <modules>
     <module>aot</module>
     <module>aot-maven-plugin</module>
+    <module>aot-runtime</module>
     <module>aot-tests</module>
     <module>bom</module>
     <module>cli</module>
@@ -138,6 +139,11 @@
       <dependency>
         <groupId>com.dylibso.chicory</groupId>
         <artifactId>aot</artifactId>
+        <version>${project.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>com.dylibso.chicory</groupId>
+        <artifactId>aot-runtime</artifactId>
         <version>${project.version}</version>
       </dependency>
       <dependency>
@@ -573,6 +579,7 @@
           <failOnWarning>${maven.dependency.failOnWarning}</failOnWarning>
           <ignoreUnusedRuntime>true</ignoreUnusedRuntime>
           <ignoredUnusedDeclaredDependencies>
+            <dependency>com.dylibso.chicory:aot-runtime</dependency>
             <dependency>com.dylibso.chicory:wasm-corpus</dependency>
             <dependency>org.junit.jupiter:junit-jupiter-engine</dependency>
           </ignoredUnusedDeclaredDependencies>
