+++ wabt-gen-aot/pom.xml
@@ -16,7 +16,7 @@
   <dependencies>
     <dependency>
       <groupId>com.dylibso.chicory</groupId>
-      <artifactId>aot</artifactId>
+      <artifactId>aot-runtime</artifactId>
     </dependency>
     <dependency>
       <groupId>com.dylibso.chicory</groupId>
